Comment part finish

+++ back-end/controllers/comment.controller.js
@@ -7,13 +7,7 @@ const User = require('../models/user');
 module.exports.getAllComments = async (req, res, next) => {
 
     const comment = await Comment.findAll({
-      // include: [
-      //   {
-      //     model: db.User,
-      //     attributes: ["firstName", "lastName"],
-      //   },
-      // ],
-      // order: [["createdAt", "DESC"]],
+      
     })
       .then((comments) => {
         res.status(200).json(comments);
@@ -26,12 +20,13 @@ module.exports.getAllComments = async (req, res, next) => {
 
 module.exports.deleteComment = async ( req, res, next) => {
     const comment = await Comment.findOne({where: { id: req.params.id} });
+    const user = await User.findOne({where: { id: req.auth} });
 
-    // if (comment.userId != req.auth ) {
-    //     return res.status(400).json({
-    //       error: new Error("Requête non autorisée"),
-    //     }),console.log('Vous navez pas les droits nécéssaires');
-    // }
+    if (comment.userId != req.auth && user.admin != true) {
+      return res.status(400).json({
+        error: new Error("Requête non autorisée"),
+      }),console.log('Vous navez pas les droits nécéssaires');
+    }
     await Comment.destroy({where : {id: req.params.id}})
       .then(() => res.status(200).json({ message: "Commentaire supprimé" }))
       .catch((error) => res.status(404).json({ error }));
@@ -40,12 +35,15 @@ module.exports.deleteComment = async ( req, res, next) => {
 module.exports.createComment = async (req, res, next) => {
     const user = await User.findOne({where: { id: req.auth} });
     delete req.body.id;
-
+    
+    console.log("req.body.id",req.body.id);
     const createComment = await Comment.create({
         ... req.body,
         description: req.body.description,
         user_name: user.user_name,
-        userId: req.auth
+        userId: req.auth,
+        postId: req.body.postId
+        
     });
   
     console.log(createComment);
